Delete cloud v1 miniapp catalog
Remove the legacy Mentra app catalog bridge that fetched and managed cloud-v1 applets through RestComms.

Keep built-in/offline app registration in a smaller catalog and let the island app store own local-only app state without host hooks.
