+++ mobile/src/components/miniapp/offlineHostedPackages.ts
@@ -4,7 +4,7 @@
  * the root expo-router stack.
  *
  * Kept separate from offlineAppRegistry so launch-decision call sites
- * (MiniappCatalog, AppSwitcher) don't transitively import every hosted
+ * (BuiltInMiniappCatalog, AppSwitcher) don't transitively import every hosted
  * screen component.
  *
  * Deliberately excludes captions / notify — no offlineRoute, nothing to render.
