+++ mobile/src/services/MantleManager.ts
@@ -9,8 +9,8 @@ import {shallow} from "zustand/shallow"
 import {bootstrapMentraJS} from "@/services/mentraJsBootstrap"
 import {phonePhotoCoordinator} from "@mentra/island"
 import {phoneStreamCoordinator} from "@mentra/island"
-import miniappCatalog from "@/services/miniapps/MiniappCatalog"
 import {preinstalledMiniappSync} from "@/services/miniapps/preinstalledMiniappSync"
+import builtInMiniappCatalog from "@/services/miniapps/BuiltInMiniappCatalog"
 import {BUNDLED_MINIAPPS} from "@/generated/bundledMiniapps"
 import {CHINA_HIDDEN_APPS, isChinaBuild} from "@/constants/miniapps"
 import {migrate} from "@/services/Migrations"
@@ -210,7 +210,7 @@ class MantleManager {
 
     // Register the offline-app catalog with island's AppRegistry before
     // anything triggers an apps refresh.
-    miniappCatalog.init()
+    builtInMiniappCatalog.init()
 
     await migrate() // do any local migrations here
     const res = await restComms.loadUserSettings() // get settings from server
